Fixing PrimeCalculator on Catapult by implementing Quartus timing constraint generation; fixing Catapult memory check
--HG--
branch : issue/HAST-131

+++ Hast.Common/Models/IDeviceManifest.cs
@@ -34,7 +34,7 @@ namespace Hast.Layer
         /// <summary>
         /// Gets the amount of memory (RAM) available to hardware implementations, in bytes.
         /// </summary>
-        uint AvailableMemoryBytes { get; }
+        ulong AvailableMemoryBytes { get; }
 
         /// <summary>
         /// The width, in bytes, of the bus available for the generated hardware implementation to exchange data with
